# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Usuarios API
  description: Ejemplo de contrato sobre usuarios fakers
  contact:
    name: Antonio Obregón
    email: antonio.obregon@utict.unam.mx
  version: 1.0.0
servers:
  - url: http://{servername}:{port}/api/{version}
    description: Servidor de pruebas
    variables:
      servername:
        # note! no enum here means it is an open value
        default: localhost
        description: Nombre del servidor API
      port:
        default: '3000'
      version:
        default: v1
tags:
  - name: users
    description: Todo sobre usuarios fakers
paths:
  /users:
    get:
      tags:
        - users
      summary: Obtiene una lista de usuarios.
      description: Devuelve una lista de objectos de usuarios disponibles.
      operationId: getUsers
      responses:
        '200':
          $ref: '#/components/responses/Usuarios.response'
        '400':
          $ref: '#/components/responses/BadRequest.response'
        '500':
          $ref: '#/components/responses/ServerError.response'
        default:
          $ref: '#/components/responses/Default.response'
    post:
      tags:
        - users
      summary: Crea un nuevo usuario
      description: Agrega un nuevo usuario
      operationId: addUser
      requestBody:
        $ref: '#/components/requestBodies/CrearUsuario'
      responses:
        '201':
          $ref: '#/components/responses/CrearUsuario.response'
        '400':
          $ref: '#/components/responses/BadRequest.response'
        '422':
          $ref: '#/components/responses/UnprocessableContent.response'
        '500':
          $ref: '#/components/responses/ServerError.response'
  /users/findById:
    get:
      tags:
        - users
      summary: Obtiene un usuario por ID.
      description: Devuelve un usuario por ID.
      operationId: findUserById
      parameters:
          - $ref: '#/components/parameters/query.parameter'
      responses:
        '200':
          $ref: '#/components/responses/Usuario.response'
        '404':
          $ref: '#/components/responses/NotFound.response'
        default:
          $ref: '#/components/responses/Default.response'
  /users/{id}:
    get:
      tags:
        - users
      summary: Obtiene un usuario por ID.
      description: Devuelve un usuario por ID.
      operationId: getUserById
      parameters:
        - $ref: '#/components/parameters/id.parameter'
      responses:
        '200':
          $ref: '#/components/responses/Usuario.response'
        '404':
          $ref: '#/components/responses/NotFound.response'
        '400':
          $ref: '#/components/responses/BadRequest.response'
        '500':
          $ref: '#/components/responses/ServerError.response'
        default:
          $ref: '#/components/responses/Default.response'
    patch:
      tags:
        - users
      summary: Actualiza porción o total de un usuario existente
      description: Actualiza una porción o el total de los datos de un usuario existente
      operationId: updateUserById
      parameters:
        - $ref: '#/components/parameters/id.parameter'
      requestBody:
        $ref: '#/components/requestBodies/ActualizarUsuario'
      responses:
        '200':
          $ref: '#/components/responses/ActualizarUsuario.response'
        '400':
          $ref: '#/components/responses/BadRequest.response'
        '404':
          $ref: '#/components/responses/NotFound.response'
        '409':
          $ref: '#/components/responses/Conflict.response'
        '422':
          $ref: '#/components/responses/UnprocessableContent.response'
        '500':
          $ref: '#/components/responses/ServerError.response'
    delete:
      tags:
        - users
      summary: Elimina un usuario por ID
      description: Elimina un usuario identificado por un ID
      operationId: deleteUserById
      parameters:
        - $ref: '#/components/parameters/id.parameter'
      responses:
        '200':
          $ref: '#/components/responses/EliminarUsuario.response'
        '404':
          $ref: '#/components/responses/NotFound.response'
        '409':
          $ref: '#/components/responses/Conflict.response'
        '422':
          $ref: '#/components/responses/UnprocessableContent.response'
        '500':
          $ref: '#/components/responses/ServerError.response'
components:
  parameters:
    id.parameter:
      name: id
      in: path
      description: ID del usuario
      required: true
      schema:
        $ref: '#/components/schemas/uuid'
    query.parameter:
      name: id
      in: query
      description: ID del usuario
      required: true
      schema:
        $ref: '#/components/schemas/uuid'
  schemas:
    uuid:
      type: string
      pattern: '^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$'
      minLength: 36
      maxLength: 36
      example: 663d5363-ea3b-43f5-be04-3e7ccef942cc
    Usuario:
      type: object
      required:
        - fullName
      properties:
        id:
          $ref: '#/components/schemas/uuid'
        fullName:
          type: string
          description: Nombre completo.
          example: 'Hal Finney'
        jobArea:
          type: string
          description: Departamento de trabajo.
          example: 'Cypherpunk'
        email:
          type: string
          format: email
          description: Correo electronico.
          example: 'Hal-Finney@yahoo.com'
        isBlock:
          type: boolean
    ErrorModel:
      type: object
      properties:
        code:
          type: string
          description: Código
        message:
          type: string
          description: Resumen del código
        description:
          type: string
          description: Descripción del código
      required:
        - code
        - message
        - description
  requestBodies:
    CrearUsuario:
      description: Datos para la creación de un usuario
      content:
        application/json:
          schema:
            allOf:
            - $ref: '#/components/schemas/Usuario'
            - type: object
              properties:
                id:
                  readOnly: true
                fullName:
                  writeOnly: true
                jobArea:
                  writeOnly: true
                email:
                  writeOnly: true
                isBlock:
                  readOnly: true
              required:
              - fullName
              - jobArea
              - email
      required: true
    ActualizarUsuario:
      description: Request body para la actualización de una porción del usuario
      content:
        application/json:
          schema:
            allOf:
            - $ref: '#/components/schemas/Usuario'
            - type: object
              properties:
                id:
                  readOnly: true
                isBlock:
                  readOnly: true
      required: true
  responses:
    Usuario.response:
      description: Usuario
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Usuario'
    Usuarios.response:
      description: Lista de usuarios
      content:
        application/json:
          schema:
            type: object
            properties:
              data:
                type: array
                items:
                  $ref: '#/components/schemas/Usuario'
    CrearUsuario.response:
      description: Usuario creado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Usuario'
    ActualizarUsuario.response:
      description: Usuario actualizado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Usuario'
    EliminarUsuario.response:
      description: Usuario eliminado
      content:
        application/json:
          schema:
            properties:
              id:
                $ref: '#/components/schemas/uuid'
    BadRequest.response:
      description: La solicitud presenta una sintaxis de incorrecta
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorModel'
          example:
            code: '400'
            message: 'Bad Request'
            description: 'Malformed request syntax'
    NotFound.response:
      description: La solicitud no obtuvo resultados en el servidor
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorModel'
          example:
            code: '404'
            message: 'Not found'
            description: 'Requested resource not found'
    Conflict.response:
      description: La solicitud presenta conflictos en el estado del modelo
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorModel'
          example:
            code: '409'
            message: 'Conflict'
            description: 'Trying to add a resource that already exists'
    UnprocessableContent.response:
      description: La solicitud presenta conflictos con alguna regla de negocio
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorModel'
          example:
            code: '422'
            message: 'Unprocessable Content'
            description: 'Business logic or Authorization errors'
    ServerError.response:
      description: El servidor encontró una condición inesperada que le impidió cumplir con la solicitud
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorModel'
          example:
            code: '500'
            message: 'Internal Server Error'
            description: 'Server encountered an unexpected condition'
    # note! Response is used for all HTTP codes that are not covered individually
    Default.response:
      description: Error inesperado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorModel'